Switch to the new rotary embedding layer
This performs the appropriate test updates and replacements for the
rotary embedding layer. This also refactors some test code to avoid
cross implementation dependencies.
