+++ sharktank/sharktank/layers/rotary_embedding.py
@@ -1,4 +1,4 @@
-# Copyright 2024 Advanced Micro Devices, Inc.
+# Copyright 2025 Advanced Micro Devices, Inc.
 #
 # Licensed under the Apache License v2.0 with LLVM Exceptions.
 # See https://llvm.org/LICENSE.txt for license information.
@@ -9,273 +9,119 @@ from typing import Optional, Union
 import torch
 
 from .base import BaseLayer
-from sharktank import ops, kernels
-from sharktank.types import (
-    ShardedTensor,
-    SplitPrimitiveTensor,
-    ReplicatedTensor,
-    ShardedTensor,
-    unbox_tensor,
-)
+
+from sharktank.kernels.mlir_kernel import *
 
 
-def build_rotary_layer(
-    tensor_parallelism_size: int = 1,
-    pipeline_parallelism: bool = False,
-    devices=None,
-    **kwargs,
-):
-    rotary_layer = RotaryEmbeddingLayer(**kwargs)
-    return ShardedRotaryLayer(
-        tensor_parallelism_size=tensor_parallelism_size,
-        pipeline_parallelism=pipeline_parallelism,
-        rotary_layer=rotary_layer,
-        devices=devices,
+def RoPEKernels():
+    BS = DynDim.BS
+    SL = DynDim.SL
+    HEADS = DynDim.HEADS
+    HALFDIM = DynDim.HALFDIM
+    TWO = StaticDim.TWO(2)
+
+    TY = Dtype.TY
+
+    @mlir_kernel(
+        inputs=(
+            MLIRTensor[BS, SL, HEADS, HALFDIM, TY],
+            MLIRTensor[BS, SL, HEADS, HALFDIM, TY],
+        ),
+        results=(MLIRTensor[BS, SL, HEADS, TWO, HALFDIM, TY],),
     )
+    def rope_select_concat(x1, x2, out=None):
+        """
+        IREE doesn't have a good concat op yet which can also do fusion. The
+        alternatives are tensor.concat or tensor.insert_slice, but both would
+        block fusions for RoPE. We use a linalg.generic with arith.select
+        on the concat dimension to do the concat instead.
+        """
+
+        mlir = """
+        !dtype = !x1_dtype
+        #trait = {
+            indexing_maps = [
+                affine_map<(bs, sl, heads, two, halfdim) -> (bs, sl, heads, halfdim)>,
+                affine_map<(bs, sl, heads, two, halfdim) -> (bs, sl, heads, halfdim)>,
+                affine_map<(bs, sl, heads, two, halfdim) -> (bs, sl, heads, two, halfdim)>
+            ],
+            iterator_types = ["parallel", "parallel", "parallel", "parallel", "parallel"]
+        }
+        module {
+        util.func private @{{kernel_name}}(%x1: !x1,
+                                           %x2: !x2) -> !out {
+            %c0 = arith.constant 0 : index
+            %c1 = arith.constant 1 : index
+            %c2 = arith.constant 2 : index
+            %c3 = arith.constant 3 : index
+
+            %bs = tensor.dim %x1, %c0 : !x1
+            %sl = tensor.dim %x1, %c1 : !x1
+            %heads = tensor.dim %x1, %c2 : !x1
+            %halfdim = tensor.dim %x1, %c3 : !x1
+            %empty = tensor.empty(%bs, %sl, %heads, %halfdim) : !out
+
+            %out = linalg.generic #trait
+                   ins(%x1, %x2 : !x1, !x2)
+                   outs(%empty: !out) {
+                ^bb0(%xs1 : !dtype, %xs2 : !dtype, %o : !dtype):
+                %two_dim = linalg.index 3 : index
+                // Ideally, when the two dim is unrolled, this condition
+                // would become a no-op and we will not do any redundant
+                // computation.
+                %is_x1 = arith.cmpi eq, %two_dim, %c0 : index
+                %val = arith.select %is_x1, %xs1, %xs2 : !dtype
+                linalg.yield %val : !dtype
+            } -> !out
+            util.return %out : !out
+        }
+        }
+        """
+        return MLIRSpec(mlir)
+
+    return rope_select_concat
 
 
-# We wrap a shardless rotary layer so the sharding behavior can be handled independently of the numerics.
-class ShardedRotaryLayer(BaseLayer):
-    def __init__(
-        self,
-        *,
-        tensor_parallelism_size: int,
-        pipeline_parallelism: bool,
-        rotary_layer,
-        devices,
-    ):
-        super().__init__()
-        self._pipeline_parallelism = pipeline_parallelism
-        self._tensor_parallelism_size = tensor_parallelism_size
-        self._rotary_layer = rotary_layer
-        self._devices = (
-            devices
-            if devices is not None
-            else tuple(range(self._tensor_parallelism_size))
-        )
-
-    def rotary_embed_table(self):
-        t = self._rotary_layer.create_rotary_embed_table()
-        if self._tensor_parallelism_size > 1 or self._pipeline_parallelism:
-            # Replicate across all devices, the data is not a lot and the computation is cheap.
-            t = ops.replicate(t, self._tensor_parallelism_size, devices=self._devices)
-
-        return t
-
-    def forward(
-        self,
-        *,
-        xt: Union[torch.Tensor, ShardedTensor],
-        start_index: int,
-    ):
-        table = self.rotary_embed_table()
-
-        if not isinstance(table, ShardedTensor):
-            return self._rotary_layer(
-                xt=xt, start_index=start_index, rotary_embed_table=table
-            )
-
-        shards = [
-            self._rotary_layer(xt=xs, start_index=start_index, rotary_embed_table=ts)
-            for xs, ts in zip(xt.shards, table.shards)
-        ]
-        return xt.clone(ts=shards)
-
-    def compute_batch_mask(
-        self, start_positions: Union[torch.Tensor, ShardedTensor], batch_seq_len: int
-    ) -> torch.Tensor:
-        if isinstance(start_positions, ShardedTensor):
-            shards = []
-            table = self.rotary_embed_table()
-            for s, ts in zip(start_positions.shards, table.shards):
-                shard = self._rotary_layer.compute_batch_mask(
-                    start_positions=s,
-                    batch_seq_len=batch_seq_len,
-                    rotary_embed_table=ts,
-                )
-                shards.append(shard)
-
-            return start_positions.clone(ts=shards)
-
-        return self._rotary_layer.compute_batch_mask(
-            start_positions=start_positions,
-            batch_seq_len=batch_seq_len,
-            rotary_embed_table=self.rotary_embed_table(),
-        )
-
-    def apply_batched_mask(
-        self,
-        *,
-        xt: Union[torch.Tensor, SplitPrimitiveTensor, ReplicatedTensor],
-        mask: Union[torch.Tensor, ReplicatedTensor],
-    ) -> Union[SplitPrimitiveTensor, ReplicatedTensor]:
-
-        if not isinstance(xt, ShardedTensor):
-            return self._rotary_layer.apply_batched_mask(xt=xt, mask=mask)
-
-        assert isinstance(mask, ReplicatedTensor) and mask.shard_count == xt.shard_count
-        xt_shards = [
-            self._rotary_layer.apply_batched_mask(
-                xt=unbox_tensor(xt_shard),
-                mask=unbox_tensor(mask_shard),
-            )
-            for xt_shard, mask_shard in zip(xt.shards, mask.shards)
-        ]
-        return xt.clone(ts=xt_shards)
+select_concat = RoPEKernels()
 
 
 class RotaryEmbeddingLayer(BaseLayer):
-    """Computes a rotary embedding in the style popularized by llama (RoPE)."""
+    """Computes a rotary embedding (RoPE)"""
 
     def __init__(
         self,
         *,
-        rope_dimension_count: int,
-        max_seqlen: int,
-        rope_freq_base: Optional[float],
-        device: Optional[torch.device] = None,
-        use_hf: bool = False,
-        use_table: bool = True,
-        dtype: torch.dtype = torch.float32,
+        head_dim: int,
+        rope_theta: float = 10000.0,
+        interleaved: bool = True,
         yarn_beta_slow: float | None = None,
         yarn_beta_fast: float | None = None,
         yarn_factor: float | None = None,
         yarn_original_context_len: int | None = None,
-        model_arch: Optional[str] = None,
     ):
         super().__init__()
-        self.device = device
-        self.rope_dimension_count = rope_dimension_count
-        self.max_seqlen = max_seqlen
-        self.use_hf = use_hf
-        self.use_table = use_table
-        self.dtype = dtype
-        self.rope_freq_base = rope_freq_base if rope_freq_base is not None else 10000.0
+        self.head_dim = head_dim
+        self.rope_theta = rope_theta
+        self.interleaved = interleaved
+
         self.yarn_beta_slow = yarn_beta_slow
         self.yarn_beta_fast = yarn_beta_fast
         self.yarn_factor = yarn_factor
         self.yarn_original_context_len = yarn_original_context_len
-        self.model_arch = model_arch
 
-    @staticmethod
-    def rotate_half(x):
-        """Rotates half the hidden dims of the input."""
-        x1 = x[..., : x.shape[-1] // 2]
-        x2 = x[..., x.shape[-1] // 2 :]
-        return torch.cat((-x2, x1), dim=-1)
-
-    def forward(
-        self,
-        *,
-        xt: torch.Tensor,
-        start_index: int,
-        rotary_embed_table: Optional[torch.Tensor],
-    ):
-        # freqs_cis shape: max_sl, dim
-        # xq_, xk_ shape: bs, sl, _, dim
-        xt_ = xt
-        _, sl, _, _ = xt_.shape
-
-        if self.model_arch == "llama4":
-            freqs_cis_real = rotary_embed_table[0][
-                :, : rotary_embed_table[0].shape[1] // 2
-            ]
-            freqs_cis_imag = rotary_embed_table[1][
-                :, : rotary_embed_table[0].shape[1] // 2
-            ]
-            # TODO: don't use complex numbers as the compiler does better without them.
-            freqs_cis = torch.view_as_complex(
-                torch.stack([freqs_cis_real, freqs_cis_imag], dim=-1)
-            )
-            freqs_cis = freqs_cis.unsqueeze(0)
-            xt_ = torch.view_as_complex(xt.float().reshape(*xt.shape[:-1], -1, 2))
-            xt_out = torch.view_as_real(xt_ * freqs_cis[:, :, None, :]).flatten(3)
-            return xt_out.type_as(xt)
-
-        if self.use_hf:
-            freqs_cis = rotary_embed_table
-            # Slice from max to current sequence length
-            cos, sin = [x[start_index : start_index + sl, :] for x in freqs_cis]
-            # expand to 1, sl, 1, dim and repeat per bs
-            cos = cos[None, :, None, :].repeat(xt.shape[0], 1, 1, 1)
-            sin = sin[None, :, None, :].repeat(xt.shape[0], 1, 1, 1)
-            xt = xt.transpose(1, 2)
-            xt_out = (xt_ * cos) + (self.rotate_half(xt_) * sin)
-            return xt_out
-
-        # Offset the table based on starting position.
-        if self.use_table:
-            freqs_cis = rotary_embed_table[start_index : start_index + sl, :]
-            freqs_cis = freqs_cis[0:sl, :]
-        else:
-            freqs_cis = torch.arange(sl, device=xt.device) + start_index
-            freqs_cis = self.compute_rotary_embed_table(freqs_cis)
-
-        assert (
-            freqs_cis.shape[0] >= sl
-        ), f"Sequence length longer than embedding table ({sl} vs {freqs_cis.shape[0]})"
-
-        freqs_cis = ops.repeat(freqs_cis[None, :, :], (xt_.shape[0], 1, 1))
-        xt_out = kernels.apply_rotary_embedding(xt_.to(freqs_cis.dtype), freqs_cis)
-
-        return ops.to(xt_out, xt.dtype)
-
-    def compute_batch_mask(
-        self,
-        start_positions: Union[torch.Tensor, ReplicatedTensor],
-        batch_seq_len: int,
-        rotary_embed_table: torch.Tensor,
-    ) -> torch.Tensor:
-        # TODO: I'm pretty sure this function is only correct because batch_seq_len is always 1
-        """Computes a mask for a batch that can be repeatedly applied.
-
-        Args:
-          start_positions: Tensor of [bs] with start positions for every sequence
-            in the batch.
-          batch_seq_len: The sequence length dimension of the batch.
-        Returns:
-          Tensor of [bs, sl, 1, d] that will be later passed to apply_batch_mask.
-        """
-        self.trace_tensor("rope.start_positions", start_positions)
-        positions_seq = torch.arange(0, batch_seq_len, device=self.device).unsqueeze(
-            0
-        ) + start_positions.unsqueeze(1)
-        # Broadcast lookup to [b, ...].
-        self.trace_tensor("rope.positions_seq", positions_seq)
-        if self.use_hf:
-            assert self.use_table, "use_hf requires use_table"
-            freqs_cis = rotary_embed_table
-            cos, sin = [x[positions_seq.flatten(), :] for x in freqs_cis]
-            freqs_cis = (cos[:, None, None, :], sin[:, None, None, :])
-            return freqs_cis
-
-        if self.use_table:
-            freqs_cis = rotary_embed_table[positions_seq.flatten()]
-        else:
-            freqs_cis = self.compute_rotary_embed_table(positions_seq.flatten())
-
-        return freqs_cis.unsqueeze(1)
-
-    def apply_batched_mask(self, *, xt: torch.Tensor, mask: torch.Tensor):
-        """Applies the embedding to a ragged batch of queries and keys.
-
-        This does a more complicated indexing operation for cases when the each
-        sequence in the batch has a potentially different start position.
-
-        positions should be of [bs, sl] and enumerate positions of all tokens.
-        """
-        # xq_, xk_ shape: bs, sl, _, dim
-        # freqs_cis shape: max_sl, dim
-
-        if self.use_hf:
-            cos, sin = mask
-            xt = xt.transpose(1, 2)
-            xt_out = (xt * cos) + (self.rotate_half(xt) * sin)
-            return xt_out.transpose(1, 2)
-
-        xt_out = kernels.apply_rotary_embedding(xt.to(mask.dtype), mask)
-
-        return xt_out.type_as(xt)
+    def _compute_theta(self):
+        # TODO: Add rope scaling.
+        dim = self.head_dim
+        # The original paper creates a d/2 dimensional space to represent
+        # the polar coordinates.
+        #
+        # From the original paper:
+        #   theta = 10000^{-2 (i - 1) / d}, i \in [1, 2, ..., d/2]
+        # which is a convoluted way of saying
+        #   theta = (1/base)^{i / d}, i \in range(0, dim, 2)
+        freqs = 1.0 / (self.rope_theta ** (torch.arange(0, dim, 2).float() / dim))
+        freqs = self._apply_yarn(freqs)
+        return freqs
 
     def _apply_yarn(self, freqs):
         yarn_factor = self.yarn_factor
@@ -314,29 +160,125 @@ class RotaryEmbeddingLayer(BaseLayer):
             freqs = torch.where(is_medium_freq, smoothed_inv_freq, inv_freq_llama)
         return freqs
 
-    def compute_rotary_embed_table(self, t):
-        dim = self.rope_dimension_count
-        if self.use_hf:
+    def compute_sincos_cache(
+        self, position_ids: torch.Tensor, dtype: torch.dtype
+    ) -> tuple[torch.Tensor, torch.Tensor]:
+        """
+        Precompute a sin/cos cache based on position_ids. This cache can
+        generally be used across. We could also rely on the compiler to do CSE,
+        but it can sometimes be really hard to do so.
 
-            freqs = 1.0 / (
-                self.rope_freq_base
-                ** (torch.arange(0, dim, 2, dtype=torch.int64).float() / dim)
-            )
-            freqs = self._apply_yarn(freqs)
-            freqs = torch.cat((freqs, freqs), dim=-1).to(device=self.device)
-            emb = t.unsqueeze(1).float() * freqs.unsqueeze(0).float()
+        position_ids: [bs, seq_len]
+        dtype: dtype for the sin/cos cache
+        device: device for the sin/cos cache
+        output: [bs, seq_len, 1, head_dim // 2], [bs, seq_len, 1, head_dim // 2]
+        """
+        theta = self._compute_theta()
+        theta_expanded = (
+            theta[None, :, None].float().expand(position_ids.shape[0], -1, 1)
+        )
+        position_ids_expanded = position_ids[:, None, :].float()
 
-            cos = torch.cos(emb).to(self.dtype)
-            sin = torch.sin(emb).to(self.dtype)
-            return (cos, sin)
+        freqs = theta_expanded @ position_ids_expanded
+        freqs = freqs.transpose(1, 2)
 
-        freqs = 1.0 / (
-            self.rope_freq_base ** ((torch.arange(0, dim) // 2).float() / dim * 2.0)
-        ).to(device=self.device)
-        freqs = self._apply_yarn(freqs)
-        freqs = (t.unsqueeze(1) * freqs.unsqueeze(0)).float()
-        return freqs
+        cos = freqs.cos().to(dtype=dtype)
+        sin = freqs.sin().to(dtype=dtype)
 
-    def create_rotary_embed_table(self):
-        t = torch.arange(self.max_seqlen, device=self.device)
-        return self.compute_rotary_embed_table(t)
+        cos = cos.unsqueeze(2)
+        sin = sin.unsqueeze(2)
+        return cos, sin
+
+    def forward(
+        self,
+        q: torch.Tensor,
+        sincos_cache: tuple[torch.Tensor, torch.Tensor],
+    ) -> tuple[torch.Tensor, torch.Tensor]:
+        """
+        Compute the rotary embedding for q/k tensors, given the sin/cos cache.
+
+        q: [bs, seq_len, heads, head_dim]
+        sincos_cache: as produced by `compute_sincos_cache`
+        output: ([bs, seq_len, heads, head_dim], [bs, seq_len, heads, head_dim])
+        """
+
+        cos, sin = sincos_cache
+        dtype = cos.dtype
+
+        def apply_rotary(x: torch.Tensor):
+            # The original RoPE paper forms "interleaved" pairs along the head
+            # dimension, i.e. it forms pairs like:
+            #   [0, 1, 2, 3, 4, 5...] -> [(0, 1), (2, 3), (4, 5), ...]
+            # and computes the embedding as:
+            #   [rope_0(0, 1), rope_1(0, 1), rope_0(2, 3), rope_1(2, 3)]
+            #
+            # This is somewhat problemetic, as this is effectively creating
+            # a innermost dimension of 2, which we are working over.
+            #
+            # However, it doesn't really matter how the pairs are formed,
+            # as long as training and inference form the same pairs. HuggingFace
+            # uses an alternative "interweaved" implementation, where pairs are
+            # formed as:
+            #   [0, 1, 2 ... 64, 65, 66...] -> [(0, 64), (1, 65), (2, 66), ...]
+            # and computes the embedding as
+            #   [rope_0(0, 64), rope_0(1, 65) ... rope_1(0, 64), rope_1(1, 65)]
+            #
+            # This implementation is prefered, because it creates an outer
+            # dimension of 2, which we can handle much better in general.
+            #
+            # Another interesting thing to note about RoPE is the output of
+            # RoPE can use either "interleaved" or "interweaved" pair layout,
+            # as long as both Q/K use the same implementation, since the
+            # computation involved is rope(Q) @ rope(K).T, where the head_dim
+            # is reduced (permutations along reduction dimensions are
+            # preserved).
+            #
+            # With this information, our implementation of RoPE supports taking
+            # both interleaved/interweaved pairing as input, however the
+            # interweaved pairing as input is prefered. If your model uses
+            # interleaved pairing, this can be converted to interweaved pairing
+            # by transposing the weights like hugging face does (hugging face
+            # uses interweaved pairing):
+            # https://github.com/huggingface/transformers/issues/25199#issuecomment-1687720247
+            #
+            # Our implementation only produces interweaved output. Note that
+            # HuggingFace also always produces the interweaved output.
+            # NOTE: Any elementwise operations along the reduction dimension
+            # post rope need to also be permuted the same way (this applies to
+            # testing equivalence checks too).
+
+            if self.interleaved:
+                # TODO: Currently, we do a strided extract slice here. This, in
+                # codegen would turn into a slow_memcpy dispatch to do this
+                # slicing. There are two ways to avoid this:
+                #   1. Use gather and fuse the gather.
+                #   2. Do expand_shape and then extract.
+                #   3. Add a deinterleave operation and fuse it.
+                #
+                # Possible issues with each of these (respectively):
+                #   1. Using gathers on the innermost dimension is effectively
+                #   taking contiguity information away. This ruins the
+                #   possibility of getting any good loads.
+                #   2. The expand_shape on the innermost dimension creates an
+                #   innermost dimension of 2. It is not the nicest to work with
+                #   when targeting intrinsics which need a dim of atleast 16.
+                #   3. We do not have a deinterleave operation in IREE as of
+                #   today.
+                #
+                # For now, we live with a slow_memcpy until someone complains.
+                x_real = x[..., 0 : self.head_dim : 2]
+                x_imag = x[..., 1 : self.head_dim : 2]
+            else:
+                x_real = x[..., : self.head_dim // 2]
+                x_imag = x[..., self.head_dim // 2 :]
+
+            x1 = x_real * cos - x_imag * sin
+            x2 = x_imag * cos + x_real * sin
+
+            cated = select_concat(x1, x2)
+            # Collapse the last two dimensions.
+            cated = cated.flatten(start_dim=-2, end_dim=-1)
+
+            return cated
+
+        return apply_rotary(q.to(dtype)).to(q.dtype)
