+++ sharktank/tests/layers/sharded_rotary_embedding_test.py
@@ -7,7 +7,7 @@
 
 import torch
 
-from sharktank.layers.rotary_embedding import build_rotary_layer
+from sharktank.layers import build_rotary_layer
 from sharktank import ops
 from sharktank.types import (
     ShardedTensor,
@@ -32,7 +32,6 @@ def test_sharded_rotary_table():
     xk = torch.rand((bs, max_seqlen, heads, rope_dims), dtype=torch.float)
     default_layer = build_rotary_layer(
         rope_dimension_count=rope_dims,
-        max_seqlen=max_seqlen,
         rope_freq_base=rope_freq_base,
     )
     oq = default_layer(xt=xq, start_index=0)
@@ -43,7 +42,6 @@ def test_sharded_rotary_table():
     xk = SplitPrimitiveTensor(ts=xk, shard_dim=2, shard_count=4)
     shard_layer = build_rotary_layer(
         rope_dimension_count=rope_dims,
-        max_seqlen=max_seqlen,
         rope_freq_base=rope_freq_base,
         tensor_parallelism_size=4,
     )
