+++ sharktank/tests/models/llama/test_llama.py
@@ -90,3 +90,4 @@ class LlamaIreeVsEagerTest(TempDirTestBase):
             iree_hal_target_device=self.iree_hal_target_device,
         )
         tester.run_and_compare_iree_vs_eager()
+
